added encryption and decryption

+++ src/main/java/com/prologics/tools/nzip/service/RecreatorService.java
@@ -5,10 +5,11 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.Base64;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
+
+import com.prologics.tools.nzip.util.EncDecUtil;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -20,35 +21,41 @@ public class RecreatorService {
             String jsonContent = Files.readString(jsonFile.toPath());
             JSONObject projectJson = new JSONObject(jsonContent);
             
-            String projectName = projectJson.getString("pname");
+            String projectName = EncDecUtil.decrypt(projectJson.getString("pname"));
             JSONArray filesArray = projectJson.getJSONArray("pcontent");
 
             File rootDirectory = new File(outputDirectory, projectName);
 
             recreateFiles(rootDirectory, filesArray);
-        } catch (IOException e) {
+        } catch (Exception e) {
             log.error("exception while converting json to project", e);
         }
     }
     
-    private void recreateFiles(File rootDirectory, JSONArray filesArray) {
+    private void recreateFiles(File rootDirectory, JSONArray filesArray) throws Exception {
     	for (int i = 0; i < filesArray.length(); i++) {
             JSONObject fileObject = filesArray.getJSONObject(i);
-            String filePath = fileObject.getString("path");
+            String filePath = EncDecUtil.decrypt(fileObject.getString("path"));
             String fileContent = fileObject.getString("content");
             createFile(rootDirectory, filePath, fileContent);
         }
     }
 
-    private void createFile(File outputDirectory, String filePath, String content) {
+    private void createFile(File outputDirectory, String filePath, String content) throws Exception {
         try {
-        	File recreatedFile = new File(outputDirectory, filePath);
-            recreatedFile.getParentFile().mkdirs();
+            File recreatedFile = new File(outputDirectory, filePath);
+            File parentDir = recreatedFile.getParentFile();
+
+            if (!parentDir.exists() && !parentDir.mkdirs()) {
+                log.error("Failed to create directories for: {}", parentDir.getAbsolutePath());
+                return;
+            }
+
             try (FileWriter writer = new FileWriter(recreatedFile)) {
-                writer.write(new String(Base64.getDecoder().decode(content)));
+                writer.write(EncDecUtil.decrypt(content));
             }
         } catch (IOException e) {
-        	log.error("exception while creating file", e);
+            log.error("exception while creating file", e);
         }
     }
 }
